src/main/java/com/infusionsoft/cas/auth/OAuthClientCredentialsTokenProvider.java
package com.infusionsoft.cas.auth;

import com.infusionsoft.cas.domain.OAuthServiceConfig;
import com.infusionsoft.cas.oauth.dto.OAuthGrantType;
import com.infusionsoft.cas.oauth.exceptions.OAuthInvalidRequestException;
import org.apache.commons.lang3.StringUtils;
import org.springframework.stereotype.Component;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

/**
 * A client credentials grant token provider
 */
@Component
public class OAuthClientCredentialsTokenProvider implements OAuthFilterTokenProvider {

    @Override
    public OAuthClientCredentialsAuthenticationToken createAuthenticationToken(HttpServletRequest request, HttpServletResponse response, String scope, String application, String grantType, OAuthServiceConfig oAuthServiceConfig, String clientId, String clientSecret) {

        if (!OAuthGrantType.CLIENT_CREDENTIALS.isValueEqual(grantType)) {
            return null;
        }

        if (StringUtils.isBlank(clientId)) {
            throw new OAuthInvalidRequestException("oauth.exception.clientId.missing");
        }
        if (StringUtils.isBlank(clientSecret)) {
            throw new OAuthInvalidRequestException("oauth.exception.clientSecret.missing");
        }

        return new OAuthClientCredentialsAuthenticationToken(null, oAuthServiceConfig, clientId, clientSecret, scope, grantType, application);
    }
}
